+++ interfaces/src/main/java/ar/edu/itba/paw/service/ProductService.java
@@ -2,17 +2,18 @@ package ar.edu.itba.paw.service;
 
 import ar.edu.itba.paw.model.Product;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 public interface ProductService {
-    Product create(int categoryId, String name, String description, byte[] image, double price);
+    Product create(int categoryId, String name, String description, byte[] image, BigDecimal price);
 
     Optional<Product> getById(int productId);
 
     List<Product> getByCategory(int categoryId);
 
-    boolean updatePrice(int productId, double price);
+    boolean updatePrice(int productId, BigDecimal price);
 
     boolean updateName(int productId, String name);
